优化动态数据源
