+++ hsweb-datasource/hsweb-datasource-jta/src/main/java/org/hswebframework/web/datasource/jta/JtaDataSourceRepository.java
@@ -5,7 +5,7 @@ package org.hswebframework.web.datasource.jta;
  *
  * @author zhouhao
  */
-public interface JtaDataSourceStore {
+public interface JtaDataSourceRepository {
 
     AtomikosDataSourceConfig getConfig(String id);
